betty style fix

+++ 0x07-pointers_arrays_strings/9-set_string.c
@@ -6,14 +6,14 @@
  */
 void set_string(char **s, char *to)
 {
-  unsigned long int i;
-  char *p;
-    
-  p = *s;
-  i = 0;
-  while (to[i] != '\0')
-    {
-      p[i] = to[i];
-      i++;
-    }
+	int i;
+	char *p;
+
+	p = *s;
+	i = 0;
+	while (to[i] != '\0')
+	{
+		p[i] = to[i];
+		i++;
+	}
 }
